+++ mk/mk/src/component/nine_patch_renderer.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 #[derive(Component)]
 pub struct NinePatchRenderer {
     pub layer: Layer,
-    pub order: isize,
+    pub order: i32,
     pub color: Color,
     pub shader: Arc<Shader>,
     pub nine_patch: Arc<SpriteNinePatch>,
@@ -14,7 +14,7 @@ pub struct NinePatchRenderer {
 impl NinePatchRenderer {
     pub fn new(
         layer: Layer,
-        order: isize,
+        order: i32,
         color: Color,
         shader: Arc<Shader>,
         nine_patch: Arc<SpriteNinePatch>,
